machinekit: use hard coded default password since rsa is now grumpy on modern os
