--- configs/pr0nscope-mk15/axis.ini
@@ -1,234 +1,0 @@
-[PRUCONF]
-DRIVER=hal_pru_generic
-# pru_period
-# 10000: default
-# 1000: cause problems, I forget what exactly
-# 2000 is fast enough for me so dont need to mess with it any more
-CONFIG=pru=0 num_stepgens=2 num_pwmgens=0 pru_period=2000
-PRUBIN=xenomai/pru_generic.bin
-
-
-[EMC]
-
-# Name of machine, for use with display, etc.
-MACHINE =               Cal-XY
-
-# Debug level, 0 means no messages. See src/emc/nml_int/emcglb.h for others
-#DEBUG =                0x00000003
-#DEBUG =                0x00000007
-DEBUG = 0
-
-
-
-
-[DISPLAY]
-
-DISPLAY = axis
-#DISPLAY = linuxcncrsh
-
-# Default: /usr/share/axis/images/axis.ngc
-# Youre supposed to be able to set to empty string to prevent load but doesnt seem to work
-# /dev/null complains about missing M30
-#OPEN_FILE=/dev/null
-
-# Cycle time, in seconds, that display will sleep between polls
-CYCLE_TIME =            0.200
-
-# Path to help file
-HELP_FILE =             tklinucnc.txt
-
-# Initial display setting for position, RELATIVE or MACHINE
-POSITION_OFFSET =       RELATIVE
-
-# Initial display setting for position, COMMANDED or ACTUAL
-POSITION_FEEDBACK =     ACTUAL
-
-# Highest value that will be allowed for feed override, 1.0 = 100%
-MAX_FEED_OVERRIDE =     1.5
-
-# Prefix to be used
-PROGRAM_PREFIX = /home/machinekit/machinekit/nc_files
-
-# Introductory graphic
-#INTRO_GRAPHIC =         machinekit.gif
-#INTRO_TIME =            5
-
-# Increments for the JOG section
-INCREMENTS = 10 1 0.1 0.01
-
-
-[FILTER]
-PROGRAM_EXTENSION = .png,.gif,.jpg Grayscale Depth Image
-PROGRAM_EXTENSION = .py Python Script
-png = image-to-gcode
-gif = image-to-gcode
-jpg = image-to-gcode
-py = python
-
-
-[TASK]
-
-# Name of task controller program, e.g., milltask
-TASK =                  milltask
-
-# Cycle time, in seconds, that task controller will sleep between polls
-CYCLE_TIME =            0.010
-
-
-
-
-[RS274NGC]
-
-# File containing interpreter variables
-PARAMETER_FILE =        pru-stepper.var
-
-
-
-
-[EMCMOT]
-
-EMCMOT =                motmod
-
-# Timeout for comm to emcmot, in seconds
-COMM_TIMEOUT =          1.0
-
-# Interval between tries to emcmot, in seconds
-COMM_WAIT =             0.010
-
-# Servo task period, in nanoseconds
-# 1000000: default
-# 10000: BBB froze and had to be restarted (suspect RT task spun)
-# 100000: "14: Unexpected realtime delay on RT thread 1"
-# 250000: lowest tested okay
-SERVO_PERIOD =          250000
-
-
-
-
-[HAL]
-
-# The run script first uses halcmd to execute any HALFILE
-# files, and then to execute any individual HALCMD commands.
-
-# list of hal config files to run through halcmd
-# files are executed in the order in which they appear
-
-HALFILE =		my.hal
-
-# list of halcmd commands to execute
-# commands are executed in the order in which they appear
-#HALCMD =               save neta
-
-
-[TRAJ]
-
-AXES =                  2
-COORDINATES =           X Y
-MAX_ANGULAR_VELOCITY = 45.00
-DEFAULT_ANGULAR_VELOCITY = 4.50
-LINEAR_UNITS =          mm
-ANGULAR_UNITS =         degree
-CYCLE_TIME =            0.010
-DEFAULT_VELOCITY =      0.50
-MAX_LINEAR_VELOCITY =   1.00
-
-
-# 51,200 steps / rev
-# 1 rev => 50 um
-# If I do mm scale is 51,200 / 0.050 = 1024000
-
-[AXIS_0]
-
-TYPE =              LINEAR
-MAX_VELOCITY =       0.225
-MAX_ACCELERATION =   0.450
-# Set Stepgen max 20% higher than the axis
-STEPGEN_MAX_VEL =    0.240
-STEPGEN_MAX_ACC =    0.480
-
-BACKLASH =           0.000
-
-# Scale: steps
-SCALE =             -1024000
-MIN_LIMIT =        -50
-MAX_LIMIT =         50
-
-FERROR =     1.0
-MIN_FERROR = 0.25
-
-HOME =                  0.000
-HOME_OFFSET =           0.00
-HOME_IGNORE_LIMITS =    YES
-HOME_USE_INDEX =        NO
-HOME_SEQUENCE =         0
-
-# Set to zero if you don't have physical home/limit switches
-# Set to the desired homing and latch velocity if you have switches
-# See: https://github.com/machinekit/machinekit-docs
-HOME_SEARCH_VEL =       0.0
-HOME_LATCH_VEL =        0.0
-
-# these are in nanoseconds
-# MDrive settings
-# Figure 6.1	 Step /direction signal timing
-DIRSETUP   =              50
-DIRHOLD    =              100
-STEPLEN    =              100
-STEPSPACE  =              100
-
-
-
-
-[AXIS_1]
-
-TYPE =              LINEAR
-MAX_VELOCITY =       0.225
-MAX_ACCELERATION =   0.450
-# Set Stepgen max 20% higher than the axis
-STEPGEN_MAX_VEL =    0.240
-STEPGEN_MAX_ACC =    0.480
-
-BACKLASH =           0.000
-
-# Scale: steps
-SCALE =             -1024000
-MIN_LIMIT =        -50
-MAX_LIMIT =         50
-
-FERROR =     1.0
-MIN_FERROR = 0.25
-
-HOME =                  0.000
-HOME_OFFSET =           0.00
-HOME_IGNORE_LIMITS =    YES
-HOME_USE_INDEX =        NO
-HOME_SEQUENCE =         0
-
-# Set to zero if you don't have physical home/limit switches
-# Set to the desired homing and latch velocity if you have switches
-# See: https://github.com/machinekit/machinekit-docs
-HOME_SEARCH_VEL =       0.0
-HOME_LATCH_VEL =        0.0
-
-# these are in nanoseconds
-# MDrive settings
-# Figure 6.1	 Step /direction signal timing
-DIRSETUP   =              50
-DIRHOLD    =              100
-STEPLEN    =              100
-STEPSPACE  =              100
-
-
-
-
-[EMCIO]
-
-# Name of IO controller program, e.g., io
-EMCIO =                 io
-
-# cycle time, in seconds
-CYCLE_TIME =            0.100
-
-# tool table file
-TOOL_TABLE =            tool.tbl
-
