--- configs/pr0nscope-mk15/microscope.json
@@ -1,101 +1,0 @@
-{
-    "live_video": true,
-    "objective_json": "objective.json",
-    "scan_json": "scan.json",
-    "cnc": {
-        "!engine":  "auto",       
-        "!engine": "mock",
-        "engine": "lcnc-arpc",
-        "!dry":true,
-        "mock":0,
-        "out_dir":"out",
-        "overwrite":0,
-        "lcnc": {
-            "host": "mk"    
-        },
-        "axes": {
-            "x": {
-                "backlash": 0.0050
-            },
-            "y": {
-                "backlash": 0.0050
-            },
-            "z": {
-                "backlash": 0.0050
-            }
-        }
-    },
-    "imager": {
-        "source":"mock",
-        "snapshot_dir":"snapshot",
-		"width": 3264,
-		"height": 2448,
-		"scalar": 0.5
-    },
-    "imager": {
-        "source":"gst-toupcamsrc",
-        "snapshot_dir":"snapshot",
-		"width": 3264,
-		"height": 2448,
-		"scalar": 0.5
-    },
-    "imager": {
-        "source":"gst-v4l2src",
-        "snapshot_dir":"snapshot",
-		"width": 3264,
-		"height": 2448,
-		"scalar": 0.5
-    },
-    "imager": {
-        "source":"gst-v4l2src-mu800",
-        "snapshot_dir":"snapshot",
-		"width": 3264,
-		"height": 2448,
-		"scalar": 0.5
-    },
-    "!imager": {
-        "source":"auto",
-        "snapshot_dir":"snapshot",
-		"width": 3264,
-		"height": 2448,
-		"scalar": 0.5
-    },
-    "objective": [
-		{
-			"name":"Mitutoyo Apo 20X",
-			"scalar": 0.5,
-			"x_view": 0.597
-		},
-		{
-			"name":"Mitutoyo Apo 2X",
-			"scalar": 0.5,
-			"x_view": 5.97
-		},
-		{
-			"name":"Mitutoyo Apo 5X",
-			"scalar": 0.5,
-			"x_view": 2.388
-		},
-		{
-			"name":"Mitutoyo Apo 10X",
-			"scalar": 0.5,
-			"x_view": 1.194
-		},
-		{
-			"name":"Mitutoyo Apo 50X",
-			"scalar": 0.5,
-			"x_view": 0.2388
-		},
-		{
-			"name":"VC 60X",
-			"scalar": 0.5,
-			"x_view": 0.199
-		},
-		{
-			"name":"Mitutoyo Apo 100X",
-			"scalar": 0.5,
-			"x_view": 0.1194
-		}
-    ]
-}
-
