--- configs/pr0nscope-mk15/my.hal
@@ -1,189 +1,0 @@
-# #######################################
-#
-# HAL file for BeagleBone + BeBoPr cape with 4 steppers
-#
-# Derived from example hm2-stepper config
-#
-# ########################################
-
-# Launch the setup script to make sure hardware setup looks good
-loadusr -w ./setup.sh
-
-
-# ###################################
-# Core EMC/HAL Loads
-# ###################################
-
-# kinematics
-loadrt trivkins
-
-# motion controller, get name and thread periods from ini file
-# trajectory planner
-loadrt tp
-loadrt [EMCMOT]EMCMOT servo_period_nsec=[EMCMOT]SERVO_PERIOD num_joints=[TRAJ]AXES tp=tp kins=trivkins
-
-
-# load low-level drivers
-loadrt hal_bb_gpio output_pins=816,822,823,824,825,826,914,923,925 input_pins=807,808,809,810,817,911,913
-loadrt [PRUCONF](DRIVER) prucode=$(HAL_RTMOD_DIR)/[PRUCONF](PRUBIN) [PRUCONF](CONFIG) halname=hpg
-loadrt limit1 count=2
-
-
-# ################################################
-# THREADS
-# ################################################
-
-addf hpg.capture-position   servo-thread
-addf bb_gpio.read                         servo-thread
-addf motion-command-handler               servo-thread
-addf motion-controller                    servo-thread
-addf limit1.0                             servo-thread
-addf limit1.1                             servo-thread
-addf hpg.update             servo-thread
-addf bb_gpio.write                        servo-thread
-
-
-# ######################################################
-# Axis-of-motion Specific Configs (not the GUI)
-# ######################################################
-
-
-# ################
-# X [0] Axis
-# ################
-
-# axis enable chain
-newsig emcmot.00.enable bit
-sets emcmot.00.enable FALSE
-
-net emcmot.00.enable <= axis.0.amp-enable-out
-net emcmot.00.enable => hpg.stepgen.00.enable
-
-
-# position command and feedback
-net emcmot.00.pos-cmd <= axis.0.motor-pos-cmd
-net emcmot.00.pos-cmd => hpg.stepgen.00.position-cmd
-
-net motor.00.pos-fb <= hpg.stepgen.00.position-fb
-net motor.00.pos-fb => axis.0.motor-pos-fb
-
-
-# timing parameters
-setp hpg.stepgen.00.dirsetup        [AXIS_0]DIRSETUP
-setp hpg.stepgen.00.dirhold         [AXIS_0]DIRHOLD
-
-setp hpg.stepgen.00.steplen         [AXIS_0]STEPLEN
-setp hpg.stepgen.00.stepspace       [AXIS_0]STEPSPACE
-
-setp hpg.stepgen.00.position-scale  [AXIS_0]SCALE
-
-setp hpg.stepgen.00.maxvel          [AXIS_0]STEPGEN_MAX_VEL
-setp hpg.stepgen.00.maxaccel        [AXIS_0]STEPGEN_MAX_ACC
-
-#setp hpg.stepgen.00.step_type       0
-# P8.43 PRU1.out2
-setp hpg.stepgen.00.steppin        813
-# P8.44 PRU1.out4
-setp hpg.stepgen.00.dirpin         812
-#setp bb_gpio.p8.out-13.invert 1
-
-
-# ################
-# Y [1] Axis
-# ################
-
-# axis enable chain
-newsig emcmot.01.enable bit
-sets emcmot.01.enable FALSE
-
-net emcmot.01.enable <= axis.1.amp-enable-out
-net emcmot.01.enable => hpg.stepgen.01.enable
-
-
-# position command and feedback
-net emcmot.01.pos-cmd <= axis.1.motor-pos-cmd
-net emcmot.01.pos-cmd => hpg.stepgen.01.position-cmd
-
-net motor.01.pos-fb <= hpg.stepgen.01.position-fb
-net motor.01.pos-fb => axis.1.motor-pos-fb
-
-
-# timing parameters
-setp hpg.stepgen.01.dirsetup        [AXIS_1]DIRSETUP
-setp hpg.stepgen.01.dirhold         [AXIS_1]DIRHOLD
-
-setp hpg.stepgen.01.steplen         [AXIS_1]STEPLEN
-setp hpg.stepgen.01.stepspace       [AXIS_1]STEPSPACE
-
-setp hpg.stepgen.01.position-scale  [AXIS_1]SCALE
-
-setp hpg.stepgen.01.maxvel          [AXIS_1]STEPGEN_MAX_VEL
-setp hpg.stepgen.01.maxaccel        [AXIS_1]STEPGEN_MAX_ACC
-
-#setp hpg.stepgen.01.step_type       0
-# P8.42 PRU1.out5
-setp hpg.stepgen.01.steppin        815
-# P8.39 PRU1.out6
-setp hpg.stepgen.01.dirpin         814
-#setp bb_gpio.p8.out-15.invert 1
-
-
-# ##################################################
-# Standard I/O - EStop, Enables, Limit Switches, Etc
-# ##################################################
-
-# Create estop signal chain
-# Drive software estop to hardware
-net estop-out iocontrol.0.user-enable-out => bb_gpio.p8.out-26
-setp bb_gpio.p8.out-26.invert 1
-
-# Monitor estop input from hardware
-net estop-loop bb_gpio.p8.in-17 => iocontrol.0.emc-enable-in
-setp bb_gpio.p8.in-17.invert 1
-
-# create signals for tool loading loopback
-net tool-prep-loop iocontrol.0.tool-prepare => iocontrol.0.tool-prepared
-net tool-change-loop iocontrol.0.tool-change => iocontrol.0.tool-changed
-
-# Axis enable signal (active low)
-net emcmot.00.enable => bb_gpio.p9.out-14
-setp bb_gpio.p9.out-14.invert 1
-
-# Machine power
-# Use halui.machine.is-on instead?
-net emcmot.00.enable => bb_gpio.p9.out-23
-
-# Tie machine power signal to the CRAMPS LED
-# Feel free to tie any other signal you like to the LED
-net emcmot.00.enable => bb_gpio.p9.out-25
-
-# ################
-# Limit Switches
-# ################
-newsig limit-x-min bit
-newsig limit-x-max bit
-newsig limit-y-min bit
-newsig limit-y-max bit
-
-net limit-x-min <= bb_gpio.p8.in-08
-net limit-x-max <= bb_gpio.p8.in-07
-net limit-y-min <= bb_gpio.p8.in-10
-net limit-y-max <= bb_gpio.p8.in-09
-
-# Adjust as needed for your switch polarity
-setp bb_gpio.p8.in-08.invert 1
-setp bb_gpio.p8.in-07.invert 1
-setp bb_gpio.p8.in-10.invert 1
-setp bb_gpio.p8.in-09.invert 1
-setp bb_gpio.p9.in-11.invert 1
-setp bb_gpio.p9.in-13.invert 1
-
-# Uncomment if you actually have limit switches setup
-# You probably want to setup homing in the INI file, as well
-#net limit-x-min => axis.0.home-sw-in
-#net limit-x-min => axis.0.neg-lim-sw-in
-#net limit-x-max => axis.0.pos-lim-sw-in
-#net limit-y-min => axis.1.home-sw-in
-#net limit-y-min => axis.1.neg-lim-sw-in
-#net limit-y-max => axis.1.pos-lim-sw-in
-
