--- configs/pr0nscope-mk20/tool.tbl
@@ -1,4 +1,0 @@
-T1 P1 D0.125000 Z+0.511000 ;1/8 end mill
-T2 P2 D0.062500 Z+0.100000 ;1/16 end mill
-T3 P3 D0.201000 Z+1.273000 ;#7 tap drill
-T99999 P99999 Z+0.100000 ;big tool number
