+++ configs/pr0nscope/microscope.json
@@ -1,8 +1,5 @@
 {
-    "live_video": true,
-    "objective_json": "objective.json",
-    "scan_json": "scan.json",
-    "cnc": {
+    "motion": {
         "engine": "lcnc-arpc",
         "mock":0,
         "out_dir":"out",
@@ -10,23 +7,8 @@
         "lcnc": {
             "host": "mk"    
         },
-        "axes": {
-            "x": {
-                "backlash": 0.0050
-            },
-            "y": {
-                "backlash": 0.0050
-            },
-            "z": {
-                "backlash": 0.0050
-            }
-        }
-    },
-    "!cnc": {
-        "engine": "mock",
-        "mock":0,
-        "out_dir":"out",
-        "overwrite":0
+        "backlash": 0.05,
+        "origin": "ul"
     },
     "imager": {
         "source":"gst-toupcamsrc",
