+++ configs/pr0nscope/my.ini
