+++ configs/pr0nscope/setup.sh
