+++ configs/pr0nscope/tool.tbl
